lista producto repo validaciones

+++ persistencia/src/main/java/co/edu/uniquindio/proyecto/service/implementacion/UsuarioServiceImpl.java
@@ -45,12 +45,23 @@ public class UsuarioServiceImpl implements UsuarioService {
     @Override
     public Usuario actualizarUsuario(Usuario usuario) throws Exception
     {
+        //busca el usuario para que valide que la cedula y email no este repetida
+        Optional<Usuario> buscado = usuarioRepo.findById(usuario.getCodigo());
+        if(buscado.isPresent()) {throw  new Exception("el codigo del usuairo ya existe");}
+
+        buscado = usuarioRepo.findByEmail(usuario.getEmail());
+        if(buscado.isPresent()) {throw  new Exception("el email del usuario ya existe");}
+
         return usuarioRepo.save(usuario);
     }
 
     @Override
     public void eliminarUsuario(int codigo) throws Exception
     {
+        //busca el usuario para que valide que la cedula si exista
+        Optional<Usuario> buscado = usuarioRepo.findById(codigo);
+        if(buscado.isEmpty()) {throw  new Exception("el codigo del usuario no existe");}
+
         usuarioRepo.deleteById(codigo);
     }
 
@@ -64,6 +75,9 @@ public class UsuarioServiceImpl implements UsuarioService {
     @Override
     public List<Producto> listaFavoritos(String correo) throws Exception
     {
+        Optional <Usuario> buscado = usuarioRepo.findByEmail(correo);
+        if(buscado.isEmpty()) {throw  new Exception("el email del usuario no existe");}
+
         return productoRepo.obtenerProductosFavoritos(correo);
     }
